Use explicit int8 cast for prediction accuracy check
Agent-Logs-Url: https://github.com/TEAT-Unicaen/R.A.I.D.-RAM-Artificial-Inspection-Dump/sessions/4ced8cd4-801b-47ce-b9d0-8a4e6e7ffc29

+++ src/train.py
@@ -71,7 +71,7 @@ def train(
             optimizer.step()
             
             total_loss += loss.item()
-            preds = (torch.sigmoid(logits) > 0.5).to(dtype=y.dtype)
+            preds = (torch.sigmoid(logits) > 0.5).to(torch.int8)
             correct += (preds == y).sum().item()
             total += y.numel()
